Add function to check if task exists and fix error handling

+++ week2/homework/src/remove.js
@@ -1,15 +1,26 @@
 const fs = require('fs');
 
+function checkIfExists(task) {
+  const textByLine = fs
+    .readFileSync('data.txt')
+    .toString()
+    .split('\n');
+
+  const exists = textByLine.filter(word => word === task);
+  if (exists == task) return true;
+  else console.log('The task you entered does not exist');
+}
+
 function remove(todoTask) {
-  if (todoTask !== undefined) {
+  if (todoTask !== undefined && checkIfExists(todoTask) == true) {
     fs.readFile('./data.txt', 'utf-8', (err, data) => {
-      if (err) throw err;
+      if (err) throw Error('File cannot be read');
 
       const newValue = data.replace(todoTask, '');
 
       fs.writeFile('./data.txt', newValue, 'utf-8', err => {
-        if (err) throw err;
-        console.log(`${todoTask} removed from the list`);
+        if (err) throw Error('File cannot be written');
+        else console.log(`${todoTask} is removed from the list`);
       });
     });
   } else {
@@ -19,7 +30,7 @@ function remove(todoTask) {
 
 function reset(file) {
   fs.writeFile(file, '', 'utf-8', err => {
-    if (err) throw err;
+    if (err) throw Error;
     console.log('File is reset');
   });
 }
